feat(i18n): implement interactive StringLocalizer and simple localization (#3804)
* feat(module: i18n): implement IStringLocalizer

* fix resource name

* add blazor string localizer

* replace localizer

* add proj

* refacor assembly

* fix namespace

* rename

* rename

* clean

* clean

* clean

* add docs

+++ site/AntDesign.Docs.Build.CLI/Command/GenerateMenuJsonCommand.cs
@@ -258,7 +258,7 @@ namespace AntDesign.Docs.Build.CLI.Command
                     {
                         [language] = new DemoMenuItem()
                         {
-                            Order = int.TryParse(docData["order"], out var order) ? order : 0,
+                            Order = float.TryParse(docData["order"], out var order) ? order : 0,
                             Title = docData["title"],
                             Url = $"docs/{segments[0]}",
                             Type = "menuItem"
